move util function to codegen internal
