+++ cpp/src/arrow/compute/kernels/codegen_internal.cc
@@ -482,6 +482,21 @@ bool HasDecimal(const std::vector<TypeHolder>& types) {
   return false;
 }
 
+void PromoteIntegerForDurationArithmetic(std::vector<TypeHolder>* types) {
+  bool has_duration = std::any_of(types->begin(), types->end(), [](const TypeHolder& t) {
+    return t.id() == Type::DURATION;
+  });
+
+  if (!has_duration) return;
+
+  // Require implicit casts to int64 to match duration's bit width
+  for (auto& type : *types) {
+    if (is_integer(type.id())) {
+      type = int64();
+    }
+  }
+}
+
 }  // namespace internal
 }  // namespace compute
 }  // namespace arrow
